Final commit, removed unused files
