+++ src/main.py
@@ -1,55 +1,44 @@
 import numpy as np
-import pandas as pd
-import matplotlib.pyplot as plt
 
-# Load the data (adjust the path as needed)
-dataframe = pd.read_csv('src/Daten_TV.txt', header=None)  # just replace with your path
-data_points = dataframe.values
-n = len(data_points)
+def konfidenz(dateiname, gamma):
+    # 1. Daten einlesen
+    A = np.loadtxt(dateiname)  # Lädt die Daten aus der Datei 'Daten_TV.txt'
+    
+    # 2. Berechnung des Mittelwerts
+    n = len(A)  # Anzahl der Datenpunkte
+    mittelwert = np.sum(A) / n
+    mu = mittelwert  # Mittelwert zur Rückgabe
+    print(f"Mittelwert (mu): {mu}")
+    
+    # 3. Berechnung der Varianz
+    varianz = np.sum((A - mittelwert) ** 2) / (n - 1)  # erwartungstreue Varianz
+    print(f"Varianz (sigma^2): {varianz}")
+    
+    # 4. Numerische Integration mit der Trapezregel (berechne Konstante c)
+    flaeche = 0.5
+    start = 0
+    ende = 0.001  # Schrittweite
+    gesucht = (gamma + 1) / 2  # Zielwert
+    
+    # Funktion für die Standardnormalverteilung
+    def standardnormalverteilung(x):
+        return (1 / np.sqrt(2 * np.pi)) * np.exp(-0.5 * x ** 2)
+    
+    # Trapezregel zur Berechnung des Integrals der Standardnormalverteilung
+    while flaeche < gesucht:
+        flaeche += (ende - start) / 2 * (standardnormalverteilung(start) + standardnormalverteilung(ende))
+        start += 0.001
+        ende += 0.001
+    
+    c = start  # das gesuchte c
+    print(f"Konstante c (aus der Trapezregel): {c}")
 
-# Calculate the mean (arithmetic mean)
-mean_value = 0
-for i in range(1, n):  # index starts at 1 in your original code
-    mean_value += data_points[i, 0]
+    # 5. Berechnung des Konfidenzintervalls
+    oben = mu + c * np.sqrt(varianz / n)
+    unten = mu - c * np.sqrt(varianz / n)
+    
+    return mu, oben, unten, varianz, c
 
-mean_value = mean_value / n  # arithmetic mean
-
-# Pass the mean to a variable mu
-mu = mean_value
-
-# Calculate the variance
-variance = data_points[0, 0]**2 - mean_value**2
-for i in range(1, n):
-    variance += data_points[i, 0]**2 - mean_value**2
-
-variance = variance / (n - 1)  # unbiased variance
-
-# Trapezoidal rule to find the value c from the standard normal distribution
-# Given density function: phi(u) = 1/sqrt(2*pi) * exp(-0.5*u^2)
-
-gamma = 0.9  # Example value for gamma; you can adjust as needed
-step_size = 0.001
-area = 0.5
-start = 0
-end = step_size
-target_area = (gamma + 1) / 2
-
-while area < target_area:
-    # Trapezoidal rule
-    area += step_size / 2 * (1 / np.sqrt(2 * np.pi) * np.exp(-0.5 * start**2) +
-                             1 / np.sqrt(2 * np.pi) * np.exp(-0.5 * end**2))
-    start += step_size
-    end += step_size
-
-# The desired c-value
-c_value = start
-print(f"The value of c is: {c_value}")
-
-# Plot the data points
-plt.figure(figsize=(10, 6))
-plt.plot(range(len(data_points)), data_points, 'ro')
-plt.xlabel('Index')
-plt.ylabel('Beautiful Values')
-plt.title("Gerky's Great and Wonderful Data")
-plt.grid()
-plt.show()
+# Testaufruf der Funktion
+mu, oben, unten, varianz, c = konfidenz('src/Daten_TV.txt', 0.99)
+print(f"Konfidenzintervall: ({unten}, {oben})")
